more fixes from crates refactor
