+++ crates/discord-types/cargo.toml
@@ -28,7 +28,7 @@ time = { version = "0.3.41", features = [
     "parsing",
     "serde-well-known",
 ] }
-chrono = { version = "0.4.41", optional = true }
+chrono = { version = "0.4.41", optional = true, features = ["serde"] }
 
 [features]
 default = ["chrono"]
