+++ crates/discord-types/src/api/endpoints/auto_moderation.rs
@@ -2,8 +2,8 @@
 
 use serde::{Deserialize, Serialize};
 
-use crate::models::api_types::auto_moderation::{AutomodAction, AutomodRule, AutomodTriggerMetadata};
-use crate::models::api_types::types::{Snowflake, Timestamp};
+use crate::api::auto_moderation::{AutomodAction, AutomodRule, AutomodTriggerMetadata};
+use crate::api::types::{Snowflake, Timestamp};
 
 /// Type: get
 ///
