+++ crates/discord-types/src/api/premium_referral.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use serde_json::Value;
 use serde_repr::{Deserialize_repr, Serialize_repr};
 
-use crate::models::{api_types::users::PartialUser, types::Timestamp};
+use crate::{api::users::PartialUser, common::Timestamp};
 
 #[derive(Serialize, Deserialize)]
 pub struct PremiumReferral {
